+++ CHANGELOG.md
@@ -17,6 +17,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added previously missing 13.6.3 and 13.6.4 to database seeder
 
+### Changed
+
+- Adjusted the displayed meta data for minor operating system versions
+- Set the alignment of icon columns to center
+
 ## [0.1.0] - 2024-03-26
 
 ### Added